Refactor CrudModal component into a new Dialogs directory for better organization. Update imports across various pages to reflect the new structure. Remove the old CrudModal implementation and enhance the ColorPicker integration in the new Dialogs/CrudModal.vue. Clean up ColorPicker.vue by removing unnecessary comments and ensuring compatibility with single color selection.
